+++ Packages/com.quickeye.one-asset/Tests/Editor/AutomaticAssetCreationTests.cs
@@ -5,12 +5,16 @@ using UnityEditor;
 
 namespace QuickEye.OneAsset.Editor.Tests
 {
+    using System;
+    using UnityEngine;
     using static TestUtils;
+
     [TestOf(typeof(OneAssetLoader))]
     [InitializeOnLoad]
     public class AutomaticAssetCreationTests
     {
-        private static bool _initializeOnLoadTestsPassed;
+        private static Exception _initializeOnLoadException;
+
         static AutomaticAssetCreationTests()
         {
             RunInitializeOnLoadTests();
@@ -18,17 +22,20 @@ namespace QuickEye.OneAsset.Editor.Tests
 
         private static void RunInitializeOnLoadTests()
         {
+            var tests = new AutomaticAssetCreationTests();
             try
             {
-                var tests = new AutomaticAssetCreationTests();
                 tests.Setup();
                 tests.Should_CreateNewAsset_When_TypeHasCreateAutomaticallyAttributeAndAssetIsMissing();
+            }
+            catch (Exception e)
+            {
+                _initializeOnLoadException = e;
+            }
+            finally
+            {
                 tests.Teardown();
                 tests.OneTimeTearDown();
-                _initializeOnLoadTestsPassed = true;
-            }
-            catch (AssertionException)
-            {
             }
         }
 
@@ -53,7 +60,10 @@ namespace QuickEye.OneAsset.Editor.Tests
         [Test]
         public void InitializeOnLoadTests()
         {
-            Assert.IsTrue(_initializeOnLoadTestsPassed);
+            if (_initializeOnLoadException == null)
+                return;
+            Debug.LogException(_initializeOnLoadException);
+            Assert.Fail("Failed");
         }
 
         [Test]
@@ -61,8 +71,13 @@ namespace QuickEye.OneAsset.Editor.Tests
         {
             var asset = OneAssetLoader.Load<SoWithCreateAutomatically>();
 
-            var assetPath = AssetDatabase.GetAssetPath(asset);
-            StringAssert.Contains(SoWithCreateAutomatically.AbsoluteAssetPath, assetPath);
+            Assert.IsNotNull(asset);
+            FileAssert.Exists(SoWithCreateAutomatically.AbsoluteAssetPathWithExtension);
+            
+            // Since Unity 6 asset database does not see assets created from InitializeOnLoad callback
+            // https://issuetracker.unity3d.com/issues/resources-dot-load-fails-to-load-assets-created-in-the-same-frame-when-called-from-a-function-with-the-initializeonloadmethod-attribute
+            // https://unity3d.atlassian.net/servicedesk/customer/portal/2/IN-77788
+            // StringAssert.Contains(SoWithCreateAutomatically.AbsoluteAssetPathWithExtension, AssetDatabase.GetAssetPath(asset));
         }
 
         [Test]
@@ -82,7 +97,7 @@ namespace QuickEye.OneAsset.Editor.Tests
             {
                 CreateAssetIfMissing = true
             };
-           
+
             var asset = OneAssetLoader.Load(options, typeof(SoWithAsset));
 
             Assert.IsTrue(AssetDatabase.Contains(asset));